dom/indexedDB/test/test_objectStore_inline_autoincrement_key_added_on_put.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Indexed Database Property Test</title>

  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript;version=1.7">

function testSteps()
{
  const IDBObjectStore = Components.interfaces.nsIIDBObjectStore;
  const IDBDatabaseException = Components.interfaces.nsIIDBDatabaseException;
  const name = window.location.pathname;
  const description = "My Test Database";

  var request = mozIndexedDB.open(name, description);
  request.onerror = errorHandler;
  request.onsuccess = grabEventAndContinueHandler;
  var event = yield;

  var db = event.target.result;

  var test = {
    name: "inline key; key generator",
    autoIncrement: true,
    storedObject: {name: "Lincoln"},
    keyName: "id",
  };

  let request = db.setVersion("1");
  request.onerror = errorHandler;
  request.onsuccess = grabEventAndContinueHandler;
  event = yield;

  let objectStore = db.createObjectStore(test.name,
                                         { keyPath: test.keyName,
                                           autoIncrement: test.autoIncrement });

  request = objectStore.add(test.storedObject);
  request.onerror = errorHandler;
  request.onsuccess = grabEventAndContinueHandler;
  event = yield;

  let id = event.target.result;
  request = objectStore.get(id);
  request.onerror = errorHandler;
  request.onsuccess = grabEventAndContinueHandler;
  event = yield;

  // Sanity check!
  is(event.target.result.name, test.storedObject.name,
     "The correct object was stored.");

  // Ensure that the id was also stored on the object.
  is(event.target.result.id, id, "The object had the id stored on it.");

  finishTest();
  yield;
}

  </script>
  <script type="text/javascript;version=1.7" src="helpers.js"></script>

</head>

<body onload="runTest();"></body>

</html>
